用namespace解决引用自己的组件的类型文件，但是还是放在了全局里，不要再类型文件里放值，会导致运行时失败
