+++ tsconfig.json
@@ -14,18 +14,16 @@
         "moduleResolution": "node",  
          "rootDir": "src",
         "baseUrl": ".", 
+         "declaration": false,
          "paths": {
               "@js/*": ["./src/js/*"],
               "@css/*": ["./src/css/*"]
          },
         "lib": ["es2015","dom"],
         "types":["node"],
-         "typeRoots" : ["./typings"]
+        "typeRoots" : ["./typings"]
      },
       "exclude": [
-        "node_modules",
-        "**/*.spec.ts"
+        "node_modules"
       ]
-    
-    
 }
