+++ typings/test.d.ts
@@ -1,0 +1,6 @@
+
+
+  export interface Test{
+      name:string;
+  }
+
